+++ opentelemetry-sdk/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## vNext
 
+## v0.21.0
+
 ### Added
 
 - Log warning if two instruments have the same name with different (#1266)